Revert "[Heliocentric] Change Projects to Support VS15 (#194)" (#195)
This reverts commit 7534b0834838f1c494399fbbbbbf184fe311a88c.

+++ Heliocentric/Core/attack.cpp
@@ -49,7 +49,7 @@ void Attack::doProjectileTravel() {
 
 void Attack::doDamage(AttackableGameObject* target) {
 	/* Let's apply instantaneous damage */
-	target->take_damage((int) damage);
+	target->take_damage(damage);
 	changeState(COOLDOWN);
 }
